+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## Unreleased
 
+* Python 2 support is removed ([#31])
+* Python 3.8 support is added
 * Smatch version is taken from git tag (see [#22][])
 * Add CI/CD configuration (see PR [#23][])
 * Better handle deinversion of special roles ([#10][])
@@ -167,4 +169,5 @@ The following are taken from an old `update_log` file:
 [#22]: https://github.com/snowblink14/smatch/issues/22
 [#23]: https://github.com/snowblink14/smatch/pull/23
 [#27]: https://github.com/snowblink14/smatch/pull/27
+[#31]: https://github.com/snowblink14/smatch/pull/31