+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * Better handle deinversion of special roles ([#10][])
 * Get `smatch-table.py` working again (part of PR [#27][])
 * Add `tests/` subdirectory and `test_top.py` ([#25])
-* Make TOP relation an attribute of the top variable ([#25])
+* Make TOP relation's value a constant string and not the top node's
+  concept to avoid double-penalizing different top concepts ([#25])
 
 ## [1.0.2][]